* [MERGE] from Test-Smoke-1.18.xx [   460] Log: * [FIX]                * Test::Smoke::Syncer::Copy->sync() didn't clean up the                  destination directory [   461] Log: * [FIX]                * The Test::Smoke::Syncer::Copy->sync() destination                  directory cleanup revealed a bug in the testsuite.
git-svn-id: https://source.test-smoke.org/svn/Test-Smoke@463 f9e6db9c-56bf-0310-a82a-abbfb51efdf0

+++ t/syncer_copy.t
@@ -32,9 +32,9 @@ SETUP: {
     local *DIR;
     opendir DIR, 'win32' or plan skip_all => "Cannot opendir 'win32': $!";
     while ( my $file = readdir DIR ) {
-        -f $file or next;
-      File::Copy::copy( File::Spec->catfile( 'win32', $file ),
-                        File::Spec->catfile( $subdir, $file ) );
+        -f File::Spec->catfile('win32', $file ) or next;
+        File::Copy::copy( File::Spec->catfile( 'win32', $file ),
+                          File::Spec->catfile( $subdir, $file ) );
     }
     closedir DIR;
     # Create a '.patch'
@@ -88,7 +88,9 @@ SKIP: {
 }
 
 END { 
-    rmtree( $ddir );
-    rmtree( $cdir );
+    unless ( $ENV{SMOKE_DEBUG} ) {
+      rmtree( $ddir );
+      rmtree( $cdir );
+    }
     chdir File::Spec->updir;
 }
